+++ api/lib/block-listeners/polkadot-node-subscription.js
@@ -1,4 +1,3 @@
-const _ = require('lodash')
 const { ApiPromise, WsProvider } = require('@polkadot/api')
 const {
   publishBlockAdded,
@@ -31,7 +30,6 @@ class PolkadotNodeSubscription {
     const networkSchemaName = this.network.id.replace(/-/g, '_')
     this.db = new database(config)(networkSchemaName)
     this.height = 0
-    this.block = undefined
     this.currentSessionIndex = 0
     this.currentEra = 0
     this.blockQueue = []
@@ -51,25 +49,14 @@ class PolkadotNodeSubscription {
     console.log('Polkadot initialized')
   }
 
-  async updateValidators() {
+  async updateDBValidators() {
     // const cosmosAPI = new this.CosmosApiClass(this.network, this.store)
     // let validators = await cosmosAPI.getAllValidators(this.height)
-    const validators = await this.addPopularityToValidators(
-      this.sessionValidators,
-      this.network.id
-    )
-    // const validatorMap = await this.getValidatorMap(validators)
-    this.updateDBValidatorProfiles(validators)
-    this.store.update({
-      height: this.height,
-      block: this.block,
-      validators: this.getValidatorMap(validators),
-      data: {
-        era: this.currentEra
-      }
-    })
-    this.updateValidatorsTimeout = setTimeout(async () => {
-      this.updateValidators()
+
+    this.updateDBValidatorProfiles(this.sessionValidators)
+
+    this.updateDBValidatorsTimeout = setTimeout(async () => {
+      this.updateDBValidators()
     }, UPDATE_VALIDATORS_INTERVAL)
   }
 
@@ -155,7 +142,6 @@ class PolkadotNodeSubscription {
       })
 
       const block = await this.polkadotAPI.getBlockByHeightV2(blockHeight)
-      this.block = block
       this.enqueueAndPublishBlockAdded(block)
 
       // We dont need to fetch validators on every new block.
@@ -177,7 +163,7 @@ class PolkadotNodeSubscription {
         this.sessionValidators = sessionValidators
         // here we start the validators loop
         if (this.network.feature_validators === 'ENABLED')
-          this.updateValidators()
+          this.updateDBValidators()
 
         if (this.currentEra < era || this.currentEra === 0) {
           console.log(
@@ -209,6 +195,15 @@ class PolkadotNodeSubscription {
         }
       }
 
+      await this.store.update({
+        height: blockHeight,
+        block,
+        validators: this.sessionValidators,
+        data: {
+          era: this.currentEra
+        }
+      })
+
       // For each transaction listed in a block we extract the relevant addresses. This is published to the network.
       // A GraphQL resolver is listening for these messages and sends the
       // transaction to each subscribed user.
@@ -236,28 +231,6 @@ class PolkadotNodeSubscription {
       console.error(`newBlockHandler failed`, error.message)
       Sentry.captureException(error)
     }
-  }
-
-  async addPopularityToValidators(validators, networkId) {
-    // popularity is actually the number of views of a validator on their page
-    const validatorPopularity = await this.db.getValidatorsViews(networkId)
-    return validators.map((validator) => {
-      const thisValidatorPopularity = validatorPopularity.find(
-        ({ operator_address }) => operator_address === validator.operatorAddress
-      )
-      // we add the popularity field to the validator
-      return {
-        ...validator,
-        popularity: thisValidatorPopularity
-          ? thisValidatorPopularity.requests
-          : 0
-      }
-    })
-  }
-
-  getValidatorMap(validators) {
-    const validatorMap = _.keyBy(validators, 'operatorAddress')
-    return validatorMap
   }
 
   // this adds all the validator addresses to the database so we can easily check in the database which ones have an image and which ones don't
